set fixture scope=function in order to release objects after each test
